Use newDefaultInstance for the different Java XML Factories
Sometimes there are other XML Factories available and some of those that are older versions cause issues in the parsing of BPMN / CMMN files.
Sometimes they cutoff some strings and we miss some data using it.
When using the default instance we can be sure that what is used by Flowable in testing is what it will be used for our users as well