Stabilise tests by checking for the async and timer jobs in one transaction
It happens that if there is a timer the first check is done in one transaction and returns nothing.
However, in the meantime the timer was moved to async, which would mean that the timers count would return empty as well
and thus stop the execution.
However, not everything has been executed.

Also adjust the Process Engine MultiInstanceTest to only run the async jobs, since we are testing the creation of a timer job when something fails