Upgrade to Spring Boot 4 and Spring Framework 7
Changes done:
* Use Jackson2 for our own DispatcherServlet
* Use a copy of the JacksonAutoConfiguration from SpringBoot 3 until we fully migrate to Jackson 3
* Test servers are not pausable
* Add certain Spring Boot specific dependencies (e.g. spring-boot-kafka, spring-boot-restclient, etc.).
They are needed since Spring Boot 4 changed how their modules look like and the auto configurations are no longer in one single jar.
Now there are JARs based on the integration.
* Rewrite the Spring Kafka retry mechanism since it got migrated from Spring Retry to the Spring Core retry.
Remove uniform random backoff. The reason for this is the fact that Spring Kafka itself no longer supports it.
Remove exponential random backoff. This would be replaced with a jitter, this is a plus / minus value with some randomness that gets applied to the delay.
* Replace deprecated Spring Security AntPathRequestMatcher with the new PathPatternRequestMatcher

Upgraded versions:
* Apache Http Client 5 - 5.5
* Jakarta Enterprise CDI - 4.1.0
* Jakarta Enterprise Concurrent - 3.1.1
* Jakarta Persistence - 3.2.0
* Jakarta Servlet - 6.1.0
* Jetty - 12.1.1
* HikariCP - 7.0.2
* Reactor Netty - 1.3.0-M7
* Spring AMQP - 4.0.0-M5
* Spring Boot - 4.0.0-M3
* Spring Kafka - 4.0.0-M5
* Spring LDAP - 4.0.0-M3
* Spring Framework - 7.0.0-M3
* Spring Security - 7.0.0-M3

Upgraded JDBC Drivers:

* DB2 - 12.1.2.0
* HSQL - 2.7.3
* MariaDB - 3.5.6
* MS SQL - 13.2.0.jre11
* Oracle - 23.9.0.25.07

Upgraded Mave Plugins:
* Maven Surefire - 3.5.4
* Liquibase - 4.33.0